Restructure teams, welcome, resources
