+++ src/client/app/admin/authentication/index.ts
@@ -5,6 +5,8 @@ export * from './signin.component';
 export * from './unauthorized.component';
 export * from './audit/user-authentication-audit.component';
 
+export * from './credentials.class';
+
 export * from './authentication.service';
 export * from './user-state.service';
