+++ src/client/app/admin/end-user-agreement/index.ts
@@ -4,6 +4,8 @@ export * from './admin-edit-eua.component';
 export * from './user-eua.component';
 export * from './audit/eua-audit.component';
 
+export * from './eua.class';
+
 export * from './eua.service';
 
 export * from './admin-eua.module';
