+++ src/client/app/admin/index.ts
@@ -3,6 +3,8 @@ export * from './admin.component';
 export * from './admin.service';
 export * from './users.service';
 
+export * from './user.class';
+
 export * from './admin.module';
 export * from './admin-routing.module';
