+++ src/client/app/audit/index.ts
@@ -2,6 +2,8 @@ export * from './audit-view-change.component';
 export * from './audit.component';
 export * from './audit-object.component';
 
+export * from './audit.classes';
+
 export * from './audit.service';
 
 export * from './audit.module';
