+++ src/client/app/help/index.ts
@@ -1,5 +1,7 @@
 export * from './help.component';
 
+export * from './help.class';
+
 export * from './help.service';
 
 export * from './help.module';
