+++ src/client/app/messages/index.ts
@@ -1,6 +1,8 @@
 export * from './messages.component';
 export * from './audit/message-audit.component';
 
+export * from './message.class';
+
 export * from './message-handler.service';
 export * from './message.service';
