+++ src/client/app/notifications/index.ts
@@ -1,5 +1,7 @@
 export * from './notification-indicator.component';
 
+export * from './notification.class';
+
 export * from './notification-handler.service';
 export * from './notification.service';
 export * from './notification-state-service';
