+++ src/client/app/resources/index.ts
@@ -1,3 +1,6 @@
 export * from './manage-resource-metadata.component';
 
+export * from './owner.class';
+export * from './resource.class';
+
 export * from './resources.module';
