+++ src/client/app/teams/manage-team.component.ts
@@ -6,11 +6,12 @@ import { Response } from '@angular/http';
 import * as _ from 'lodash';
 import { Observable } from 'rxjs/Observable';
 
+import { AlertService } from 'app/shared';
+import { ConfigService } from 'app/core';
+
 import { Team } from './teams.class';
 import { TeamsService } from './teams.service';
-import { AlertService } from 'app/shared/alert.service';
 import { AuthenticationService } from 'app/admin/authentication/authentication.service';
-import { ConfigService } from 'app/core/config.service';
 
 @Component({
 	selector: 'manage-team',
