+++ src/client/app/teams/tags/audit/tag-audit.component.ts
@@ -1,7 +1,6 @@
 import { Component } from '@angular/core';
 
-import { AuditObjectTypes } from 'app/audit/audit.classes';
-import { DefaultAudit } from 'app/audit/audit-object.component';
+import { AuditObjectTypes, DefaultAudit } from 'app/audit';
 
 @Component({
 	selector: 'tag-audit',
