+++ src/client/app/teams/tags/index.ts
@@ -2,6 +2,8 @@ export * from './list-tags.component';
 export * from './manage-tag.component';
 export * from './audit/tag-audit.component';
 
+export * from './tags.class';
+
 export * from './tags.service';
 
 export * from './tags.module';
