+++ src/client/app/teams/team-summary.component.ts
@@ -5,9 +5,10 @@ import { Router, ActivatedRoute, Params } from '@angular/router';
 import * as _ from 'lodash';
 import { Modal } from 'angular2-modal/plugins/bootstrap';
 
+import { AlertService } from 'app/shared';
+
 import { Team, TeamMember } from './teams.class';
 import { TeamsService } from './teams.service';
-import { AlertService } from 'app/shared/alert.service';
 import { AuthenticationService } from 'app/admin/authentication/authentication.service';
 
 @Component({
