+++ src/client/app/teams/teams-routing.module.ts
@@ -1,7 +1,8 @@
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 
-import { AuthGuard } from 'app/core/auth-guard.service';
+import { AuthGuard } from 'app/core';
+
 import { ListTeamsComponent } from './list-teams.component';
 import { ManageTeamComponent } from './manage-team.component';
 import { TeamSummaryComponent } from './team-summary.component';
