+++ src/client/app/welcome/welcome-routing.module.ts
@@ -1,7 +1,8 @@
 import { NgModule }     from '@angular/core';
 import { RouterModule } from '@angular/router';
 
-import { AuthGuard } from 'app/core/auth-guard.service';
+import { AuthGuard } from 'app/core';
+
 import { WelcomeComponent } from './welcome.component';
 
 @NgModule({
