+++ src/client/app/welcome/welcome.module.ts
@@ -2,11 +2,12 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { Ng2BootstrapModule } from 'ng2-bootstrap';
 
+import { UtilModule } from 'app/shared';
+import { MessagesModule } from 'app/messages';
+
 import { WelcomeComponent } from './welcome.component';
 import { ExternalLinksComponent } from './external-links.component';
-import { UtilModule } from 'app/shared/util.module';
 import { WelcomeRoutingModule } from './welcome-routing.module';
-import { MessagesModule } from 'app/messages/messages.module';
 
 @NgModule({
 	imports: [
